Move read tests into io suite
